From f0cdd43f231d8eb56173a9f0beac0b86ad614f51 Mon Sep 17 00:00:00 2001 Message-ID: <f0cdd43f231d8eb56173a9f0beac0b86ad614f51.1748515590.git.gang.ji@cloud.com> [PATCH] xen: Add new XEN_DOMCTL_getvcpurunnabletime hypercall
Now the runnable time of a vCPU can't be queried.

Add domctrl hypercall XEN_DOMCTL_getvcpurunnabletime to allow querying
the runnable time of a vCPU.

And in libxc, add xc_get_vcpu_runnable_list to allow querying runnable
time of all the vCPUs of a domain, it calls domctrl hypercall:
getvcpurunnabletime for each vCPU in the domain within a multicall
hypercall.

A binding for Ocaml is also added: domain_get_vcpu_runnable_list.

Signed-off-by: Gang Ji <gang.ji@cloud.com>